+++ src/engine/atoms/index.ts
@@ -40,5 +40,4 @@ import './询问闪';
 import './询问杀';
 import './请求回应';
 import './选将';
-import './武圣包装';
-import './武圣还原';
+